rename galois prime to be more generic in the future
